actualizacion

+++ laravel/app/controllers/FirmaController.php
@@ -7,26 +7,41 @@ class FirmaController extends \BaseController
         if ( Request::ajax() ) {
             $pagina      =   Input::get('pagina');
 
-            DB::beginTransaction();
             $paginas=PaginaFirma::find($pagina);
-            $paginas['estado']=2;
-            $paginas->save();
+            if( count($paginas)>0 ){
+                if($paginas->estado==2){
+                    $aParametro['msj'] = "Página no existe";
+                    $aParametro['pagina']=$paginas->id;
+                    $aParametro['rst'] = 2;
+                }
+                else{
+                    DB::beginTransaction();
+                    $paginas['estado']=2;
+                    $paginas->save();
 
-            $delete1='  DELETE FROM escalafon_fichas_recepcion 
-                        WHERE escalafon_ficha_id IN 
-                            (SELECT id 
-                            FROM escalafon_fichas 
-                            WHERE escalafon_id='.$paginas->escalafon_id.')';
-            $delete2='DELETE FROM escalafon_fichas WHERE escalafon_id='.$paginas->escalafon_id;
-            $delete3='DELETE FROM fichas WHERE hoja='.$paginas->id;
-            DB::delete($delete1);
-            DB::delete($delete2);
-            DB::delete($delete3);
-            
-            DB::commit();
-            $aParametro['msj'] = "Se realizó la eliminación";
-            $aParametro['pagina']=$paginas->id;
-            $aParametro['rst'] = 1;
+
+                    $delete1='  DELETE FROM escalafon_fichas_recepcion 
+                                WHERE escalafon_ficha_id IN 
+                                    (SELECT id 
+                                    FROM escalafon_fichas 
+                                    WHERE escalafon_id='.$paginas->escalafon_id.')';
+                    $delete2='DELETE FROM escalafon_fichas WHERE escalafon_id='.$paginas->escalafon_id;
+                    $delete3='DELETE FROM fichas WHERE hoja='.$paginas->id;
+                    DB::delete($delete1);
+                    DB::delete($delete2);
+                    DB::delete($delete3);
+                    
+                    DB::commit();
+                    $aParametro['msj'] = "Se realizó la eliminación de la página nro ".$pagina;
+                    $aParametro['pagina']=$paginas->id;
+                    $aParametro['rst'] = 1;
+                }
+            }
+            else{
+                $aParametro['msj'] = "Página no existe";
+                $aParametro['pagina']=0;
+                $aParametro['rst'] = 2;
+            }
 
             return Response::json($aParametro);
         }
